Crate buck2_wrapper_common

Code shared between buck2_wrapper and buck2.

Careful! The wrapper is not released as part of the regular buck version bumps, meaning code changes here are not “atomically” updated.

Modules

invocation_id

is_buck2

kill

Cross-platform process killing.

pid

Constants

BUCK2_WRAPPER_ENV_VAR

BUCK_WRAPPER_START_TIME_ENV_VAR

BUCK_WRAPPER_UUID_ENV_VAR

DOT_BUCKCONFIG_D

EXPERIMENTS_FILENAME

Functions

killall

Kills all running Buck2 processes, except this process’s hierarchy. Returns whether it succeeded without errors.